ref: #518 move validate
