+++ src/ffmpeg/compile/validate.py
@@ -12,12 +12,12 @@ from __future__ import annotations
 
 from dataclasses import replace
 
-from ..compile.context import DAGContext
+from ..dag.nodes import FilterNode, InputNode
+from ..dag.schema import Node, Stream
 from ..exceptions import FFMpegValueError
 from ..streams.audio import AudioStream
 from ..streams.video import VideoStream
-from .nodes import FilterNode, InputNode
-from .schema import Node, Stream
+from .context import DAGContext
 
 
 def remove_split(
